pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "annotation-pipeline"
version = "0.1.0"
description = "VIA annotation pipeline for Mask R-CNN defect detection"
requires-python = ">=3.9"
dependencies = [
    "pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "numpy>=1.24.0",
]

[project.scripts]
annotation-pipeline = "annotation_pipeline.cli:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
include = ["annotation_pipeline*"]
